Move Abstract*Test class files to test directory
