+++ nupic/test/abstract_temporal_memory_test.py
